+++ FreeMarket/src/main/java/com/freemarket/freemarket/product/application/ProductManagementService.java
@@ -3,7 +3,6 @@ package com.freemarket.freemarket.product.application;
 import com.freemarket.freemarket.global.email.exception.EmailVerificationException;
 import com.freemarket.freemarket.global.file.application.FileStorageService;
 import com.freemarket.freemarket.product.api.dto.ProductDto;
-import com.freemarket.freemarket.product.api.dto.ProductWithStatsDto;
 import com.freemarket.freemarket.product.domain.*;
 import com.freemarket.freemarket.product.exception.ProductException;
 import com.freemarket.freemarket.user.domain.User;
@@ -12,8 +11,6 @@ import com.freemarket.freemarket.user.exception.UserException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.coyote.BadRequestException;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -25,19 +22,15 @@ import java.util.List;
 @Service
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
-public class ProductService {
-
+public class ProductManagementService {
     private final ProductRepository productRepository;
     private final UserRepository userRepository;
     private final FileStorageService fileStorageService;
-    private final ProductViewService viewService;
-    private final ProductWishlistService wishlistService;
 
     // 상품 등록
     @Transactional
     public ProductDto.ProductBaseResponse createProduct(Long userId, ProductDto.CreateRequest request, List<MultipartFile> images) throws BadRequestException {
-        User seller = userRepository.findById(userId)
-                .orElseThrow(() -> new UserException.UserNotFoundException(userId));
+        User seller = getUser(userId);
 
         // 학교 이메일 인증 여부 확인
         emailVerification(seller);
@@ -83,9 +76,10 @@ public class ProductService {
     }
 
     // 상품 수정
+
     @Transactional
     public ProductDto.ProductBaseResponse updateProduct(Long userId, Long productId, ProductDto.UpdateRequest request,
-                                                    List<MultipartFile> newImages, List<Long> deleteImageIds) {
+                                                        List<MultipartFile> newImages, List<Long> deleteImageIds) {
         Product product = getProductWithSellerCheck(productId, userId);
         product.update(request.name(), request.description(), request.price(), request.stock(), request.category());
         if (request.status() != null) product.changeStatus(request.status());
@@ -140,78 +134,19 @@ public class ProductService {
             throw new ProductException.ProductUpdateException(e.getMessage());
         }
     }
-
-    // 상품 단건 조회
-    public ProductDto.ProductDetailResponse getProduct(Long productId, Long userId) {
-        Product product = findProduct(productId);
-
-        // 조회수
-        Long viewCount = viewService.getViewCount(productId);
-        // 관심 등록 수
-        Long wishlistCount = wishlistService.getWishlistCount(productId);
-
-        // 현재 사용자의 관심 등록 여부
-        boolean isWishlisted = userId != null ? wishlistService.isWishlisted(userId, productId) : false;
-        return ProductDto.ProductDetailResponse.from(product, viewCount, wishlistCount, isWishlisted);
-    }
-
-    // 상품 목록 조회
-    public Page<ProductDto.ProductDetailResponse> getProducts(String keyword, ProductStatus status, Pageable pageable, Long userId) {
-        return productRepository.findAllWithStatsAndWishlist(keyword, status, pageable, userId)
-                .map(ProductWithStatsDto::toProductDetailResponse);
-    }
-
-    // 카테고리별 상품 조회
-    public Page<ProductDto.ProductDetailResponse> getProductsByCategory(ProductCategory category, String keyword, ProductStatus status,
-                                                                  Pageable pageable, Long userId) {
-        return productRepository.findByCategoryWithStatsAndWishlist(category, keyword, status, pageable, userId)
-                .map(ProductWithStatsDto::toProductDetailResponse);
-    }
-
     // 상품 삭제
+
     @Transactional
     public void deleteProduct(Long userId, Long productId) {
         Product product = getProductWithSellerCheck(productId, userId);
         productRepository.delete(product);
         log.info("상품 삭제 완료: 상품 ID {}, 판매자 ID {}", productId, userId);
     }
-
-    // 판매 완료 처리
-    @Transactional
-    public ProductDto.ProductBaseResponse markProductAsSold(Long sellerId, Long productId, Long buyerId) {
-        Product product = getProductWithSellerCheck(productId, sellerId);
-
-        if (product.getStatus() == ProductStatus.SOLD_OUT) {
-            throw new ProductException.AlreadySoldProductException();
-        }
-
-        User buyer = userRepository.findById(buyerId)
-                .orElseThrow(() -> new UserException.UserNotFoundException(buyerId));
-
-        product.markAsSold(buyer);
-
-        log.info("상품 판매완료 처리: 상품 ID {}, 판매자 ID {}, 구매자 ID {}", productId, sellerId, buyerId);
-
-        return ProductDto.ProductBaseResponse.from(product);
-    }
-
-    // 판매완료 취소 처리 메서드
-    @Transactional
-    public ProductDto.ProductBaseResponse cancelProductSold(Long sellerId, Long productId) {
-        Product product = getProductWithSellerCheck(productId, sellerId);
-
-        if (product.getStatus() != ProductStatus.SOLD_OUT) {
-            throw new ProductException.NotSoldProductException();
-        }
-
-        product.cancelSold();
-        log.info("판매완료 취소 처리: 상품 ID {}, 판매자 ID {}", productId, sellerId);
-        return ProductDto.ProductBaseResponse.from(product);
-    }
-
     // 판매자 권한 확인 후 상품 조회
-    private Product getProductWithSellerCheck(Long productId, Long userId) {
-        Product product = findProduct(productId);
+
+    Product getProductWithSellerCheck(Long productId, Long userId) {
+        Product product = productRepository.findById(productId)
+                .orElseThrow(() -> new ProductException.ProductNotFoundException(productId));
 
         if (!product.getSeller().getId().equals(userId)) {
             throw new ProductException.ProductAccessDeniedException();
@@ -219,17 +154,14 @@ public class ProductService {
 
         return product;
     }
-    
-    private Product findProduct(Long productId) {
-        Product product = productRepository.findById(productId)
-                .orElseThrow(() -> new ProductException.ProductNotFoundException(productId));
-        return product;
-    }
-
     private static void emailVerification(User seller) {
         if (!seller.isEmailVerified()) {
             throw new EmailVerificationException.EmailNotVerifiedException();
         }
     }
 
+    private User getUser(Long userId) {
+        return userRepository.findById(userId)
+                .orElseThrow(() -> new UserException.UserNotFoundException(userId));
+    }
 }
